mysite/requirement.txt
amadeus==3.1.0
darkskylib==0.3.91
Django==2.2
django-tempus-dominus==5.1.2.4
geopy==1.19.0
